successfully implemented velocity modulators to hyper bounces
changed ints to floats to use with the random function.

randomly generates values to add to both the ball's vx and vy values following a hyper smash, increasing speed.

resets speed to default on goal; the direction is determined by whoever scored.

vx is now random on startup.
vy is now random on both reset AND startup.