changed reward system
before, you "pulled" your territory towards your side to cover the entire screen. i've inverted it to a "push" system, which rewards the winning player even more by giving gradually less room for the losing player (the one on "defense", so to speak) to anticipate from which angle the ball will spawn.
my first idea was to "balance" the game by giving the losing player a bigger chance at a comeback, but after a bit of testing, i've found it to be much more fun and intense this way.
the next step will be to implement a proper game over screen with the option to restart, with perhaps a title screen as well.
more ambitious ideas include a sort of ultra-rudimentary "character selection", which would change the colours of the paddle and the hyper stocks, as well as substantially altering hyper smash behaviour.
or, different "types" of hyper smashes could be randomly rewarded: the red one, as already implemented, remains as a "power smash", increasing the velocity values by a random amount. a blue one, for example, could alter the velocities mid-flight, making for some curve-ball type action. a green hyper-smash could potentially hinder your opponent, for example by temporarily shortening their paddle, or potentially lowering their paddle speed. perhaps even you could hold on to 1 of each type, and each would have a specific key to press, allowing for some deliberate combo action (eg. hinder your opponent then go for a power smash).
other considerations include sound effects and image-based graphics but i rather like the minimalist aesthetic, so we'll see.