finally fixed the logic problems
one issue i've noticed is that i still get an error in the console on the first goal.
i am guessing that it is because of the check to see which song is playing in order to stop it - but since it is the first goal, no song can be playing, so it gets confused.