implemented Parasite into main code, but program no longer works
the error is that on the final line of code, parasites[p].display(), in which it states the variable "p" does not exist
i'm trying to figure out why it thinks p does not exist, as the code to create the parasites is nearly identical to the code to create the griddies. will commence troubleshooting.