added comments; got rid of modulators i didn't want
clearer explanation of previous commit: the strength at which Cells gravitate towards brightestPixel is twice as strong as the strength at which they gravitate towards the mouse's position.
try placing the cursor where brightestPixel is and watch as all the cells converge.
will probably add a small degree of randomness to these variables to get a less uniform convergence.