serious problem encountered with parasite hitbox
i've run into a serious issue regarding the range at which Parasites attack Cells.
for the sake of debugging, i've made it so that cells do not lose energy when moving. i've also made it so that when a Cell encounters a Parasite, it dies instantly.
you will notice how the hitboxes (ie. the Parasite attack range) remains fixed at the program startup. ie., the hitbox does not follow the Parasite object, it remains in a static location, and any Cells that pass over that location die.
i've been trying to figure out why this is, as the logic for the Parasite attack() function is similar to the Antibody heal() function, yet the Antibodies are functioning normally.