hitbox problem solved
there was a problem with the math that has been fixed.
i have yet to revert the values to non-debugging values, therefore cells are still insta-killed when they come in range of a parasite.
the next problem seems to be exponentially growing parasites.
i'm trying to get them to stop growing after a single kill by constraining a Cell's energy to 0 (if a parasite lowers its energy to a value below 0), checking to see if a Cell's energy == 0, and then lowering that cell's energy if it is.
there must be a problem in the logic somewhere. will check.