back to predictable areas of behaviour, except for parasite growth
the only thing that doesn't function is parasite growth, which is strange, because i had gotten it to work before.
i would guess that there is a logic step overriding the part of the parasite code that raises the killCount and hitbox variables (which are the variables used to modify the Parasites' sizes as well as their attack range), but i am at a loss at how this is the case because Parasites are still killing Cells just fine. the code to modify the size & hitbox is just after a constrain function that brings a killed cell's energy to 0 right before lowering it to a negative value.
something tells me i might have to rewrite the entire collision logic, but i'm so close...