after a bit of testing, it appears parasite hitboxes are completely non-functional, again, despite having made no major changes to the logic.
this is extremely frustrating, because i don't even know what is wrong at this point. i haven't known what was wrong the last 5 commits or something, and it keeps changing for better or worse. i'm going to try reverting to an older version but i feel like i'm just about done with this at this point.