started exercise 07
i am starting with my template completed in exercise05 as i was very upset with the mess i had made with exercise06.
on that note, i may give up on camera interaction altogether, as i was unable to generate anything that didn't end up making the game less frustrating NOR more fun. thus, i will lower the priority of incorporating camera elements. it is the effectiveness/enjoyability of the final product that i am interested in, and the camera seemed to detract from that rather than contribute to it.

the goal for this assignment is to incorporate sound, which i have already [rudimentarily] done, however i will try experimenting with [rudimentary] generative music aspects, adding original sound effects to certain events that occur in-game (ie. cells dying, cells being attacked/healed). if the sounds are to be of a musical nature, the decision of key/scale/aesthetic is needed.

a concern i have before even beginning is the busy-ness of the sound. A LOT of events are occuring at any given moment, and i wish to avoid aurally overwhelming the player with a sonically indecipherable mess. thus i anticipate having to tune at least two things: 1. lowering the overall speed of the game 2. lowering the amount of cells (which may necessitate other aspects being changed, such as their resilience, etc...)

as a gameplay aside, i am considering reverting from the bouncing physics achieved in exercise06 back to the wrapping behaviour of objects. i found that, despite being more interesting from a CODING perspective, tuning the bouncing behaviour physics to be MORE FUN was proving to be more difficult than is ultimately worth it for the enjoyability of the game. aesthetically, as well, it simple FEELS much better than constraining the objects. i will keep the bouncing code and may try to incorporate it later on when more gameplay aspects are implemented (such as huddling of cells).