wrote code for huddling, but program stalls on mouseClicked.
thought this would work fine, but the game totally freezes on a mouse click.
will troubleshoot, but what's weird is there is no error message or crash, and the reset function (the 'r' key) ceases to function as well. i can't even close the window with CMD+W, the whole thing just locks up and the only way to get rid of it is clicking the red X.

i have commented out the problematic code in the Energy class and will look into it.