fixed globalKillCount variable
game now ends properly when all 200 cells are dead

ISSUE: the killCount variable for the parasite seems to go crazy. once cells start dying, the rate at which it increases goes up at an exponential rate, like, 100s per tick kind of thing. the original intention was to count Cells whose energy = 0 as "killed", and then lowering their energy to -1, so that they cannot be "killed again". however, this does not seem to be functioning properly and so i will look into it.