found the killCount issue but having problems with solution
the issue is the parasite counts cell corpses within its hitbox as additional kills, which is strange, because i had specifically written the code to avoid this from happening.
it's somewhat of a logical mess at the moment but the basic gist of it is this: the parasite's killCount is supposed to go up by 1 when an overlapping Cell's energy = 0. on the next Cell.update cycle, it checks to see if a Cell's energy = 0, and if it is, it lowers it by 1 (making it -1, and drawing the red X), which i thought would prevent the "if (host.energy == 0)" logic from running, as the cell's energy is not supposed to be 0 anymore, it is supposed to be -1.
i'm guessing is that there is a rogue constrain function somewhere that bumps the value back up to 0 before the parasite attack function is run, so i will check it out.