haha! fixed. kill count now increases as intended.
rewrote a bit of the logic. the main issue was in the parasite.attack function; constrain was bringing any cell's energy level below 0 back to 0, even if it was "confirmed dead", ie. at -1. now, checks to see if host.energy is less than 0 BEFORE the kill check/constrain function, and breaks out of the loop if it is.