moved kill sound picker into the Parasite object
seems to have fixed the sound issue, but of course a new separate issue has arisen, in which Cells' deaths don't seem to be updating properly anymore: in this version, i've noticed that the red kill markers sometimes do not appear until a certain point, at which a bunch will appear at once. will look into it.