added comments, other tweaks
the herding tone now begins playing again if the '2' key is pressed while empty.
before it would continue to just play the 'empty' audio.

however, a new issue seems to have arisen in regards to the parasites (again): occasionally, it seems like a parasite's kill count jumps up unrealistically high very quickly. my hypothesis is that something is preventing the cell from being flagged as "confirmed dead", and so it continually gives kills to a single parasite. will look into it.