updated comments
game works nice and smooth and is fun to play.
will need to add sound effects for the newly added abilities (energy recharge and parasite stun).
also will need to add a splash screen explaining game concepts and controls (possible difficulty selector as well?)
final goal will be to create some kind of end-state. ideally (time permitting), i would like the player to face a "boss" of some type, but i am unsure as to how to implement it in what would make for both an interesting and fun encounter.