created final_project folder; added start function
the game does not run right away anymore; clicking on the screen in necessary to begin the game. a splash screen containing instructions/music will be added.
the 'R' key also resets the game to this state; clicking begins the game.
also a few of the recently added features, namely the energy recharge ability ('2') and stun ability ('3'), do not yet have sounds recorded.