added the ability to "feed" stunned parasites with healthy cells in order to disable them
once a parasite eats 50 cells (kept track of by the variable eatCount), it becomes disabled, as represented by a change in colour (becomes pink).
disabled cells no longer respond to other objects. not sure if i want to make the disabling permanent, which would make the objective of the game to disable all of the parasites, which could potentially trigger a boss encounter of sorts... potentially.

i've played a few games and so far have not been able to disable all the parasites, but the flow of the game is pretty nice. it becomes more and more difficult to keep alive a large number of cells as the game progresses. not sure at the moment if difficulty needs adjustment; will continue to try to "beat" the game but may lower the difficulty if too difficult. i might raise the number of cells.

still needs splash screen and additional sound effects.