created new project file (final_project_B) to work on the Boss class as this is potentially the biggest addition and potentially the biggest mess
added Boss class to final_project_B.
for testing purposes, now gives the player a hyper stock every second instead of every 3. this allows abilities to be spammed, trvializing the game's challenge, however it's fun to mess around more and this way makes it much easier to reach the point where the boss spawns (you must disable all parasites by feeding them 50 cells while stunned)

currently, the Boss is just your typical parasite, only bigger. and red.
but it spawns correctly after all parasites are disabled, so we're off to a roaring start!
haven't tested much yet, as there is not much new behaviour (yet). i will make sure the boss is killable and that all abilities function properly when applied to the boss.
this will be fun - but the ideas i have might be too lofty, so i will be taking things slowly and one small step at a time.