fixed bug
there was something bothering me about when cells died: it seemed as though when a large amount of cells died in a very small area, any other cells that would pass over that area would die extremely quickly if not instantly. i tested extensively and found this exactly to be the case. turns out a really old line in the code that i had written to prevent dead cells from interacting with each other was breaking LIVE cells out of their behaviour due to faulty logic.
removing the 'or' flag on a dead cell's collision behaviour seems to have fixed this, but more testing is required to make sure everything runs as intended.
anyways, because of this, cells die normally - ie., there aren't spontaneously generated "death zones" on the screen anymore, and thus it is MUCH easier to keep cells alive.
it's funny: i've been playing with that bug since the beginning now and kind of got used to it. it was like an additional "mechanic" so as to encourage keeping many cells from dying in one spot. on the other hand, it could work to your advantage: the more dead cells on a single spot means more efficient use of antibodies! the funny thing is that even though it was an unintended 'glitch', i am not even sure at the moment if it adds to or detracts from the gameplay.
will test more.

also working on boss tail behaviour but i can't seem to get the bezier curves to update every frame for some strange reason. the starting anchor frames (ie, the boss's XY coordinates, or the center of its 'body') update perfectly fine but everything else seems to change only every couple of seconds. weirder still is that the changes are seemingly arhythmic. probably just missing something trivial in the code but we'll see.