changed various parameters governing parasite & boss behaviour
increased the constrain range that governs a parasite's maximum size. they can grow much bigger now, and therefore are more of a threat if not taken care of swiftly.

boss tails now flail erratically when stunned; they move slow and smoothly when normal.
next would be a way to control the end coordinates (their second anchor points) of the beziers so that their positions are  corelated to the boss itself somehow and not simply semi-random points in the center of the window.

and finally added a new visual indicator for when parasites' stun duration is almost up: they vibrate with more intensity in the final 32 frames of the stun animation. i love small additions like this, they help a ton and really contribute to the feel of the game as well as benefitting the player's experience.