first attempt at implementing Minion attack behaviour
i've had a lot of trouble getting this thing to run for the past hour or so. lots of moving code around, modifying syntax, etc.
right now, the program works (ie. it doesn't crash), but all the minions spawn at the end of only one of the boss's bezier tails. must be a problem with for loop/array implementation somewhere, but i thought i'd commit now as it's been a really long time since i've gotten it running with NullPointerException showing up.