the boss's minions now must be severed before the boss's body can be damage
i don't like the visual feedback for attacking the boss, though. will change it so that it is more obvious that it is invulnerable while minions are alive, and more obvious that you are doing damage to it once they are all severed.
also might consider adding more minions. they're cool.