beginning work on intro sequence - not working
i thought this would be simple, using for loops to time the duration of each intro sequence image (intro1.png, intro2.png, etc...) but i get no image on startup.

the game runs fine after the for loops, and the old splash screen still works, but i'm just getting a blank grey window for the intro's duration, which doesn't make sense to me at the current moment.

the code is extremely simple as of now, so the problem is [probably] extremely simple as well.