how to play screen now stays
small issue: if you click in the same area as the button to escape from the how to play screen, you will just immediately jump back into it. may just ignore this as i can tell it's either going to take 4 seconds or a hell of a lot longer to solve so i'll come back to it if i have time.